requirements.txt
pyrebase4
streamlit